Move runes under edit.
