+++ edit/edit.go
@@ -8,8 +8,8 @@ import (
 	"sync"
 	"unicode"
 
+	"github.com/eaburns/T/edit/runes"
 	"github.com/eaburns/T/re1"
-	"github.com/eaburns/T/runes"
 )
 
 // A Buffer is an editable rune buffer.
